Move meta sync logic to its own file
